backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pymysql==1.1.0
openai==1.54.0
httpx==0.27.2
python-dotenv==1.0.0

